proof of concept
can handle request on port 3000 and read request content such as header. however the actual loop uses blocking fd's wich is not allowed, need to be changed later. thus once the request read the program blocks.

poll() needs to be understood to be used properly and the loop improved to handle request and send the correct response.